Allow deprecated code to be ignored
When refactoring a framework or library to improve layering, old code often
cannot be removed immediately. Instead, it has to be marked as @deprecated,
and kept around for a relase or two. When that is the case, we generally
want to be able to ignore violations triggered by deprecated code, since
we already know that we will remove it as soon as possible, and we want
to focus on making the new code clean. Sometimes it's not even possible
to avoid the offending dependency in the deprecated code, for structural
reasons - that may well have been the reason for deprecating that code
in the first place.

The detection logic for deprecated methods and classes is built directly
into FilweReferenceVisitor. That's not great, but serves as a proof of
concept.