Add "must_any" to BoolCollector
Use case: Assume we have a layer X that consists of classes that have either
"Impl" in the class name, or that habe the "@impl" tag. It uses two
collectors, and any class matching either one will be part of the layer.

No we want to exclude things that are in layer Y from layer X. This is
not possible with the available "must" and "must_not" options in the
BoolCollector, since we have no way to keep the idea that classes with
"Impl" in the name OR the "@impl" tag should be part of the layer
(unless they are already part of layer Y).

The introduction of the "must_any" attribute allows to address this use
case.