Configure layer internal tag (#1318)
* Expose all tags on TokenReference

Collect all @tags from doc blocks and exposes them on
ClassLikeReference and FunctionReference objects.

* Introduce TagValueRegexCollector

This collector allows to define a layer based on doc-block tags on
classes and functions. The definition can include a regular expression
that the tag value must match.

* Make psalm happy

* Use list<string> instead of string[]

* remove unused getTagText

* Psalm: remove redundant baseline entries

* Convert TaggedReferenceTrait.php into a base class

* Make @internal tag configurable

Some projects, like MediaWiki, use the @internal tag with a different
semantics than the one imposed by deptrac. In the case of MediaWiki, the
intended semantics is "not for use by extensions", i.e. "internal to the
core repo".

To work around this issue, allow the config file to specify which tag
deptrac will interpret as "layer-internal". The @deptrac-internal tag
will continue to always work.

NOTE: This is a breaking change: the "@internal" tag will will now be
ignored unless explicitly configured to be used.

* Follow up to review

- fix documentation
- use explicit null check

* Fix nitpicks

* Introduce Analyser config object

Analyser is defined analogous to Layer.

This adds a phpunit test for the relevant parts of DeptracConfig.

* Remove redundant null check in Analyser config.

* Rename Analyser class to AnalyserConfig

Also: in our own deptrac config, set internal_tag to "@internal",
since that's the tag that is used in a couple of places.

* cs fix

---------

Co-authored-by: Gennadi McKelvey <gennadijanzen@icloud.com>