Fixes: #1381
Fixes a case where wi silently skip over collectors that we deem "unresolvable". This is a vestigial case from time when we first implemented "Layer" collector. With the current state of the code, it is not necessary to check if a collector is "resolvable" and it can just fail when it throws an exception while we try to actually run it.

These exceptions are already properly covered and rethrown to the end-user.

With this change the result of running the deptrac.yaml from the issue will cause:

```console

 [ERROR] Analysis finished with an Exception.

         Invalid collector definition.

         Unknown layer "Domain_typo" specified in collector.
```